[lumen] allow LLVMDialect to inherit LLVMContext
By allowing the LLVMDialect to be created with a reference
to a non-owned LLVMContext, we can create one LLVMContext
per thread and use it for everything. This prevents the issue
where MLIR-generated modules were created with their own context
and cannot be linked together or worked with using the thread-global
LLVMContext we create for that purpose in Lumen.